added common check to simplify
